+++ src/main/java/co/programmers/voucher_management/voucher/controller/VoucherCommandLineRunner.java
@@ -1,4 +1,4 @@
-package co.programmers.voucher.controller;
+package co.programmers.voucher_management.voucher.controller;
 
 import java.io.IOException;
 import java.text.MessageFormat;
@@ -8,12 +8,12 @@ import org.slf4j.LoggerFactory;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Controller;
 
-import co.programmers.voucher.dto.Response;
-import co.programmers.voucher.dto.VoucherRequestDTO;
-import co.programmers.voucher.service.CreationService;
-import co.programmers.voucher.service.InquiryService;
-import co.programmers.voucher.view.InputView;
-import co.programmers.voucher.view.OutputView;
+import co.programmers.voucher_management.voucher.dto.Response;
+import co.programmers.voucher_management.voucher.dto.VoucherRequestDTO;
+import co.programmers.voucher_management.voucher.service.CreationService;
+import co.programmers.voucher_management.voucher.service.InquiryService;
+import co.programmers.voucher_management.view.InputView;
+import co.programmers.voucher_management.view.OutputView;
 
 @Controller
 public class VoucherCommandLineRunner implements CommandLineRunner {
