+++ src/main/java/co/programmers/voucher_management/voucher/dto/VoucherResponseDTO.java
@@ -1,4 +1,4 @@
-package co.programmers.voucher.dto;
+package co.programmers.voucher_management.voucher.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,8 +7,8 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 @Builder
-public class VoucherRequestDTO {
+public class VoucherResponseDTO {
+	private int id;
 	private String discountType;
 	private Integer discountAmount;
-	private String discountStrategy;
 }
