+++ src/main/java/co/programmers/voucher_management/voucher/repository/VoucherRepository.java
@@ -1,10 +1,10 @@
-package co.programmers.voucher.repository;
+package co.programmers.voucher_management.voucher.repository;
 
 import java.io.IOException;
 import java.util.List;
 
-import co.programmers.voucher.dto.VoucherResponseDTO;
-import co.programmers.voucher.entity.Voucher;
+import co.programmers.voucher_management.voucher.dto.VoucherResponseDTO;
+import co.programmers.voucher_management.voucher.entity.Voucher;
 
 public interface VoucherRepository {
